+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.3.0
+
+* Add RentSummary override attributes for bedroom and bathroom counts
+```
+#bedrooms_override_from_unit_type
+#bathrooms_override_from_unit_type
+```
+
+If these are not set in the initialization hash, they are computed from the unit_type string as follows:
+```
+rs = RentSummary.new(unit_type: "4.5x2.5")
+rs.bedrooms_override_from_unit_type
+=> 4.5
+rs.bathrooms_override_from_unit_type
+=> 2.5
+```
+
 ## 0.2.0 (2012-01-16)
 
 * Moved the client_name to the configuration, since it is required